% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjReturns.R
\name{adjReturns}
\alias{adjReturns}
\title{Calculates the adjusted returns}
\usage{
adjReturns(detrendReturn, paramList)
}
\arguments{
\item{detrendReturn}{TODO}

\item{paramList}{TODO}
}
\value{
adjusted returns
}
\description{
\code{adjReturns} is a function that calculates the adjusted returns
for a given data-frame with returns.
}
\examples{
detrendReturn <- data.frame("A" = rnorm(12) / 10, "B" = rnorm(12) / 10)
direction <- c(1, -1)
sharpeAdj <- c(0.9, 1.2)
volAdj    <- c(1.1, 0.9)
frequency <- 12
paramList <- list(direction, sharpeAdj, volAdj, frequency)
myAdjRets <- adjReturns(detrendReturn, paramList)
}

