% Generated by roxygen2 (4.0.2): do not edit by hand
\name{annFreq}
\alias{annFreq}
\title{Calculates the factor to annualize returns.}
\usage{
annFreq(dates)
}
\arguments{
\item{dates}{an object with dates}
}
\value{
annualization factor
}
\description{
\code{annFreq} is a function that calculates the factor to annulize returns.
Given an object with dates it determines the median difference between
consecutive dates and determines the factor to annulize return series.
}
\examples{
dates  <- seq(as.POSIXct("2010-01-01"), by="month", length.out=20)
myFreq <- annFreq(dates)
}

