% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asserta}
\alias{asserta}
\title{Assert a value against a predicate.}
\usage{
asserta(x, predicate, msg = "Error", type = c("stop", "warn", "info",
  "none"))
}
\arguments{
\item{x}{The input argument.}

\item{predicate}{The predicate function.}

\item{msg}{The message to be passed to the function in case the
assertion fails.}

\item{type}{The action to be performed in case that the assertion
fails.}
}
\value{
The original input argument if the procedure is not
    stopped due to an assertion failure.
}
\description{
Checks the value against a predicate, takes one of the four
available actions (see \code{type} argument), and return the
original value if the procedure is not \code{stop}ped.
}
\details{
TODO: Try to optimise this.
}

