% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{asVector}
\alias{asVector}
\title{Convenience Vector Cast}
\usage{
asVector(x, names = NULL, fill = NA, type = c("numeric", "logical",
  "integer", "complex", "character"))
}
\arguments{
\item{x}{The R object.}

\item{names}{Name vector if any.}

\item{fill}{The value to replace NAs with.}

\item{type}{The type to cast to.}
}
\value{
New vector to be returned.
}
\description{
Provides a convenience function to cast, name and NA-treat a given
vector compatible R object.
}

