% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JohnsonSU.R
\name{bisectionRoot}
\alias{bisectionRoot}
\title{Performs a bisection root-finding.}
\usage{
bisectionRoot(func, lower, upper, tol = 1e-06, maxiter = 10000)
}
\arguments{
\item{func}{TODO}

\item{lower}{TODO}

\item{upper}{TODO}

\item{tol}{TODO}

\item{maxiter}{TODO}
}
\value{
The bisection root.
}
\description{
Performs a bisection root-finding.
}

