% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify}
\alias{classify}
\title{Decorate an object with class}
\usage{
classify(x, klass)
}
\arguments{
\item{x}{Object to be classified.}

\item{klass}{The class name.}
}
\value{
Classified object.
}
\description{
Appends a new class name to an object's class vector and returns.
}

