% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{effectify}
\alias{effectify}
\title{Find Effective Constraints}
\usage{
effectify(directions, assetMin, assetMax, minNet, maxNet, minGrs, maxGrs)
}
\arguments{
\item{directions}{Asset directions.}

\item{assetMin}{Minimum possible absolute asset-wise weights.}

\item{assetMax}{Maximum possible absolute asset-wise weights.}

\item{minNet}{The minimum total net weight.}

\item{maxNet}{The maximum total net weight.}

\item{minGrs}{The minimum total gross weight.}

\item{maxGrs}{The maximum total gross weight.}
}
\value{
A list of weights and success status.
}
\description{
Finds effective portfolio constraints.
}
\examples{
## Successful:
effectify(rep(1, 10), rep(0, 10), rep(0.5, 10), -1, 1, 0, 2)

## Unsuccessful:
effectify(rep(1, 10), rep(0.4, 10), rep(0.5, 10), -1, 1, 0, 2)

}

