% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JohnsonSU.R
\name{johnsonSim}
\alias{johnsonSim}
\title{Simulates the distribution based on Johnson SU paramters}
\usage{
johnsonSim(mean, sd, skewness, excess.kurtosis, sims = 1000)
}
\arguments{
\item{mean}{TODO}

\item{sd}{TODO}

\item{skewness}{TODO}

\item{excess.kurtosis}{TODO}

\item{sims}{TODO}
}
\value{
The Johnson SU distribution parameters.
}
\description{
Simulates the distribution based on Johnson SU paramters
}

