% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/JohnsonSU.R
\name{johnsonSU}
\alias{johnsonSU}
\title{Computes the Johnson SU distribution parameters.}
\usage{
johnsonSU(mean, sd, skewness, excess.kurtosis)
}
\arguments{
\item{mean}{TODO}

\item{sd}{TODO}

\item{skewness}{TODO}

\item{excess.kurtosis}{TODO}
}
\value{
The Johnson SU distribution parameters.
}
\description{
Computes the Johnson SU distribution parameters.
}

