% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bangwa.R
\name{mvReturns}
\alias{mvReturns}
\title{Computes simple returns of multivariate level series per asset.}
\arguments{
\item{levels}{A numeric matrix representing multivariate level
series per column per asset.}
}
\value{
A numeric matrix representing multivariate simple returns
    per column per asset.
}
\description{
Computes simple returns of multivariate level series per asset.
}

