% Generated by roxygen2 (4.0.2): do not edit by hand
\name{performanceAnalytics}
\alias{performanceAnalytics}
\title{\code{performanceAnalytics} TODO}
\usage{
performanceAnalytics(x, frequency)
}
\arguments{
\item{x}{TODO}

\item{frequency}{TODO}
}
\description{
\code{performanceAnalytics} TODO
}
\examples{
myIndex   <- seq(as.POSIXct("2000-01-01"), as.POSIXct("2012-01-01"), by="month")
data      <- as.xts(cbind(rnorm(length(myIndex)), rnorm(length(myIndex))), myIndex)
pAnalytix <- performanceAnalytics(data[,1], 12)
}

