% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/portRandomizer.R
\name{portRandomizer}
\alias{portRandomizer}
\title{Constructs \code{orsDEoptim} parameters.}
\usage{
portRandomizer(weights, minWeight, maxWeight, minTotalNetWeight,
  minTotalGrossWeight, maxTotalNetWeight, maxTotalGrossWeight, noOfSteps, sims)
}
\arguments{
\item{weights}{A vector of weights as provided by the PM.}

\item{minWeight}{The minimum possible weight constraint per asset.}

\item{maxWeight}{The maximum possible weight constraint per asset.}

\item{minTotalNetWeight}{The minimum possible total net weight
constraint per portfolio.}

\item{minTotalGrossWeight}{The minimum possible total gross weight
constraint per portfolio.}

\item{maxTotalNetWeight}{The maximum possible total net weight
constraint, ie. sum of all position weights.}

\item{maxTotalGrossWeight}{The maximum possible total gross weight
constraint, ie. sum of all absolute position weights.}

\item{noOfSteps}{The number of steps within the min/max
boundaries. The greater the finer, thus more simulated deviations.}

\item{sims}{TODO}
}
\value{
A list of random deviations, portfolios, lower bound,
and upper bound.
}
\description{
Constructs \code{orsDEoptim} parameters.
}
\examples{
weights             <- c(0.10, -0.05, 0.20, 0.05, -0.01)
minWeight           <- 0.01
maxWeight           <- 0.25
minTotalNetWeight   <- -1
minTotalGrossWeight <- 0
maxTotalNetWeight   <- 1
maxTotalGrossWeight <- 1.5
noOfSteps           <- 10
sims                <- 10000
randPorts <- portRandomizer(weights, minWeight, maxWeight, minTotalNetWeight, minTotalGrossWeight, maxTotalNetWeight, maxTotalGrossWeight, noOfSteps, sims)
}

