% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/portSelecter.R
\name{portSelecter}
\alias{portSelecter}
\title{Selects a portfolio from a set of portfolios}
\usage{
portSelecter(priceSeries, startDate, endDate, weights, lookBack, minWeight,
  maxWeight, minTotalNetWeight, minTotalGrossWeight, maxTotalNetWeight,
  maxTotalGrossWeight, noOfSteps, sims, statsFN, selectFN, dataFN = NULL, ...)
}
\arguments{
\item{priceSeries}{A xts data-frame with asset prices.}

\item{startDate}{TODO}

\item{endDate}{TODO}

\item{weights}{TODO}

\item{lookBack}{TODO}

\item{minWeight}{TODO}

\item{maxWeight}{TODO}

\item{minTotalNetWeight}{TODO}

\item{minTotalGrossWeight}{TODO}

\item{maxTotalNetWeight}{TODO}

\item{maxTotalGrossWeight}{TODO}

\item{noOfSteps}{TODO}

\item{sims}{TODO}

\item{statsFN}{TODO}

\item{selectFN}{TODO}

\item{dataFN}{TODO}

\item{...}{TODO}
}
\value{
A list of random deviations, portfolios, lower bound,
and upper bound.
}
\description{
Selects a portfolio from a set of portfolios
}

