% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simEvent.R
\name{simPortDirections}
\alias{simPortDirections}
\title{TODO}
\usage{
simPortDirections(hitRatio, returns)
}
\arguments{
\item{hitRatio}{TODO}

\item{returns}{TODO}
}
\value{
simulated directions.
}
\description{
TODO
}

