% Generated by roxygen2 (4.0.2): do not edit by hand
\name{treatPriceSeries}
\alias{treatPriceSeries}
\title{This function removes NA's and optionally detects and remove outliers.}
\usage{
treatPriceSeries(price, outlierTreatment = FALSE, OHLC = FALSE,
  quantile = 0.998, surpressPause = FALSE)
}
\arguments{
\item{price}{The xts price series (either univariate or OHLC).}

\item{outlierTreatment}{Flag for treating outliers.}

\item{OHLC}{Flag to indicate whether we are dealing with OHLC.}

\item{quantile}{The quantile for the outlier detection}

\item{surpressPause}{TODO}
}
\value{
Cleaned price series
}
\description{
This function removes NA's and optionally detects and remove outliers.
}

