% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bangwa.R
\name{ttbd}
\alias{ttbd}
\title{Simulate Target-Mean Bounded Values}
\arguments{
\item{target}{The target mean.}

\item{lower}{A vector of lower bounds.}

\item{upper}{A vector of upper bounds.}
}
\value{
A vector of simulated values.
}
\description{
Simulates a vector of target-mean bounded values for the given target
mean, lower and upper bounds.
}
\details{
The length of the simulated values depends on the length of lower
and upper bound vectors.
}

