% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bangwa.R
\name{uvReturns}
\alias{uvReturns}
\title{Computes simple returns of univariate level series of an asset.}
\arguments{
\item{levels}{A numeric vector representing univariate level
series of an asset.}
}
\value{
A numeric vector representing unitvariate simple returns
    of an asset.
}
\description{
Computes simple returns of univariate level series of an asset.
}

