% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify}
\alias{classify}
\title{Create a Named List with Class Attribute}
\usage{
classify(..klass, ...)
}
\arguments{
\item{..klass}{The class name(s).}

\item{...}{Arbitrary named properties.}
}
\value{
Classified object.
}
\description{
Create a Named List with Class Attribute
}

