% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getPublicData.R
\name{getPublicQuote}
\alias{getPublicQuote}
\title{Constructs \code{orsDEoptim} parameters.}
\usage{
getPublicQuote(tickers, startDate, endDate, quote, provider, compression)
}
\arguments{
\item{tickers}{TODO}

\item{startDate}{TODO}

\item{endDate}{TODO}

\item{quote}{TODO}

\item{provider}{TODO}

\item{compression}{TODO}
}
\value{
List with OHLC data.
}
\description{
Constructs \code{orsDEoptim} parameters.
}

