% Generated by roxygen2 (4.0.2): do not edit by hand
\name{outlierPlot}
\alias{outlierPlot}
\title{This function plots the original price series and the corresponding
outlier treated series.}
\usage{
outlierPlot(originalSeries, newSeries, outlierIndex)
}
\arguments{
\item{originalSeries}{TODO}

\item{newSeries}{TODO}

\item{outlierIndex}{TODO}
}
\description{
This function plots the original price series and the corresponding
outlier treated series.
}

