% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sampleDates}
\alias{sampleDates}
\title{A function which returns a list with each element showing chungked sample dates.}
\usage{
sampleDates(data, startDate, endDate, lookBack, timeZone = "GMT",
  format = "\%Y-\%m-\%d")
}
\arguments{
\item{data}{TODO}

\item{startDate}{TODO}

\item{endDate}{TODO}

\item{lookBack}{TODO}

\item{timeZone}{TODO}

\item{format}{TODO}
}
\description{
\code{sampleDates} is a function that returns a list with each element of
the list showing the in-sample start date, in-sample end date and
out-of-sample date
}
\examples{
myIndex   <- seq(as.POSIXct("2000-01-01"), as.POSIXct("2012-01-01"), by="month")
data      <- as.xts(cbind(rnorm(length(myIndex)), rnorm(length(myIndex))), myIndex)
startDate <- "2005-01-01"
endDate   <- "2008-01-01"
lookBack  <- 36
mySamples <- sampleDates(data, startDate, endDate, lookBack)
}

