% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simEvent}
\alias{simEvent}
\title{Simulates a binary events.}
\usage{
simEvent(eventProb, length.out, noSim, nonEvent = -1)
}
\arguments{
\item{eventProb}{a numeric vector with the probabilities of event. Length equals the number of variables.}

\item{length.out}{the number of rows for each variable}

\item{noSim}{the number of simulations for each set}

\item{nonEvent}{the value for the non-events}
}
\value{
simulated binary event
}
\description{
\code{simEvent} TODO
}
\examples{
eventProb <- c(0.55, 0.6)
length.out<- 200
noSim     <- 1000
nonEvent  <- -1
myEvents  <- simEvent(eventProb, length.out, noSim, nonEvent)
}

